requirements.txt
google-cloud-vision
requests
opencv-python
speechrecognition
librosa
numpy
scikit-learn
soundfile